Streamlit_WebApp/requirements.txt
streamlit==1.10.0
pandas==1.1.2
numpy==1.19.5
joblib==0.16.0
Pillow==7.2.0